requirements.txt
fastapi
greenlet
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
pytest
httpx
pytest-asyncio
prometheus-client
starlette-prometheus
